Add files via upload
This is just setting up a fictitious database to build a credit reporting program on. There is a python 3 script for generating random 9 digit SSNs as well as Regular payments, and another Python 3 script for generating random phone numbers. Everytime these scripts run the results are a new set of random outcomes. Excel and word are used. After using google and bing to get some city names, common first names, common last names, common street names, and manually search for each zipcode, then the same for area codes. Save as csv scripts for those columns that need to be joined together. Like the area code and the random 7 digit phone number generator. Open it up in word, find and replace the ',' with '' a space and then Ctrl+a, Ctrl+c, then go to excel and insert a new field for entire phone number and use the mouse to insert the clipboard phone number to the excel field. Then for the addresses, use the street name, with an adjacent field of random numbers, just type something in and use the scroll crossbar when highlighted down the length of the column. Then save as a csv file, open in Word. This time find and replace the ',' with a space ' '. so that you have random street addresses. When ECOA is 1 there is no 2nd co-borrower needed. Wipe those fields as in the FakeClientDbECOA.xlsx file. The other fields have been added, and some calculated fields that show up in the excel xlsx file not the csv form. Sort by last name to mix up the files. This is entirely fake, and later a program will be created for the Tax ID numbers that are the SSNs beginning with 9. and calculating the DPD, AmtPastDue, DOFD, Repossessions, ... when adding these instances manually and adjusting those picked observations to show the late fees and days past due or charge off dates and amount charged off, and assigning a numeric or letter account status and numeric or letter payment rating. This repository prepares the dataset to be entered into a Metro 2 software program to generate the required fields to send to the three major Credit Bureaus on a monthly cycle for installment loans for automobiles.